refactor: 키워드 선택 최대 개수 변경
